Refactor CPU correlator

+++ tests/test_extensions.py
@@ -9,7 +9,7 @@ class TestExtensions(TestCase):
 
     def test_rotate_grid3d(self):
 
-        grid = np.zeros((4, 5, 6), dtype=np.float64)
+        grid = np.zeros((4, 5, 6), dtype=np.float32)
         grid[0, 0, 0] = 1
         grid[0, 0, 1] = 1
         grid[0, 1, 1] = 1
@@ -17,14 +17,14 @@ class TestExtensions(TestCase):
         grid[0, 0, -1] = 1
         grid[-1, 0, 0] = 1
         # Identity rotation
-        rotmat = np.asarray([[1, 0, 0], [0, 1, 0], [0, 0, 1]], dtype=np.float64)
+        rotmat = np.asarray([[1, 0, 0], [0, 1, 0], [0, 0, 1]], dtype=np.float32)
         out = np.zeros_like(grid)
         rotate_grid3d(grid, rotmat, 2, out, True)
         self.assertTrue(np.allclose(out, grid))
 
         # 90 degree rotation around Z-axis
         out.fill(0)
-        rotmat = np.asarray([[0, -1, 0], [1, 0, 0], [0, 0, 1]], dtype=np.float64)
+        rotmat = np.asarray([[0, -1, 0], [1, 0, 0], [0, 0, 1]], dtype=np.float32)
         rotate_grid3d(grid, rotmat, 2, out, False)
         # Build answer
         answer = np.zeros_like(out)
